Organize main app code into components and domain
