+++ src/app/components/number.tsx
@@ -1,0 +1,6 @@
+import { numFormatter } from '../domain/number-format'
+
+export function Number({ value }: { value: number }) {
+    const numFormat = numFormatter.format(Math.round(value))
+    return <span>{numFormat}</span>
+}
