+++ src/app/domain/items/config.ts
