+++ src/app/domain/items/generate.ts
