+++ src/app/domain/store.ts
@@ -1,10 +1,9 @@
-import { AppAction, RootState } from "./state";
-import { combineReducers, configureStore } from "@reduxjs/toolkit";
-import { createEpicMiddleware } from "redux-observable";
-import { electronCountReducer } from "./electrons";
-import { itemReducers } from "./items/state";
-import { gameTickEpic } from "./items/generate";
-
+import { AppAction, RootState } from './state'
+import { combineReducers, configureStore } from '@reduxjs/toolkit'
+import { createEpicMiddleware } from 'redux-observable'
+import { electronCountReducer } from './electrons'
+import { itemReducers } from './items/state'
+import { gameTickEpic } from './items/generate'
 
 const epicMiddleware = createEpicMiddleware<AppAction, AppAction, RootState>()
 
@@ -21,4 +20,3 @@ epicMiddleware.run(gameTickEpic)
 export type AppState = RootState
 export type AppDispatch = typeof store.dispatch
 export type AppStore = typeof store
-
