+++ src/app/index.tsx
@@ -1,12 +1,12 @@
 import { createRoot } from 'react-dom/client'
-import { Provider as ReduxProvider } from "react-redux";
-import { store } from "./store";
+import { Provider as ReduxProvider } from 'react-redux'
+import { store } from './domain/store'
 import { App } from './app'
 
 const element = document.getElementById('root')
 const root = createRoot(element)
-root.render((
+root.render(
     <ReduxProvider store={store}>
-        <App/>
+        <App />
     </ReduxProvider>
-))
+)
